Dev 11.3

+++ CyRK/cy/common.hpp
@@ -4,10 +4,10 @@
 #include <limits>
 
 // Pre-processor constants
-static const int Y_LIMIT      = 32;
-static const int DY_LIMIT     = 64;  // dy limit is defined by Y_LIMIT and number of extra output allowed. Typically we allow 2x the Y_LIMIT.
-static const int MESSAGE_SIZE = 128;
-static const int BUFFER_SIZE  = 16;
+static const size_t Y_LIMIT     = 32;
+static const size_t DY_LIMIT    = 64;  // dy limit is defined by Y_LIMIT and number of extra output allowed. Typically we allow 2x the Y_LIMIT.
+static const int MESSAGE_SIZE   = 128;
+static const size_t BUFFER_SIZE = 16;
 
 // Integration Constants
 // Multiply steps computed from asymptotic behaviour of errors by this.
@@ -60,14 +60,14 @@ struct MaxNumStepsOutput
 void round_to_2(size_t& initial_value);
 
 MaxNumStepsOutput find_max_num_steps(
-    const int num_y,
-    const int num_extra,
+    const size_t num_y,
+    const size_t num_extra,
     const size_t max_num_steps,
     const size_t max_ram_MB);
 
 
 size_t find_expected_size(
-    int num_y,
-    int num_extra,
+    size_t num_y,
+    size_t num_extra,
     double t_delta_abs,
     double rtol_min);
